Width and height are floating point; Scale matrix in transformation getter is now multiplied by 2 to get correct scaling

+++ src/Camera2D.hpp
@@ -18,15 +18,15 @@ class Camera2D
 {
 	public:
 		Camera2D();
-		Camera2D(Moveable2D moveable, JU::uint32 width, JU::uint32 height);
+		Camera2D(Moveable2D moveable, JU::f32 width, JU::f32 height);
 
 		void setMoveable(Moveable2D moveable);
 		void getWorld2NDCTransformation(glm::mat3& view) const;
 
 	private:
 		Moveable2D moveable_;	//!< Moveable data
-		JU::uint32 width_;		//!< Width of the camera's window
-		JU::uint32 height_;		//!< Height of the camera's window
+		JU::f32    width_;		//!< Width of the camera's window
+		JU::f32    height_;		//!< Height of the camera's window
 };
 
 } /* namespace JU */
